achitecture services updates
