+++ docker-compose.yaml
@@ -53,7 +53,7 @@ services:
       - db
     networks:
       - my_network
-
+    command: ["uvicorn", "main:app", "--reload", "--host", "0.0.0.0", "--port", "8000"]
   db:
     image: postgres:14.3
     restart: always
