+++ routers/user.py
@@ -2,27 +2,21 @@ from fastapi import APIRouter, HTTPException, Depends
 from sqlalchemy.orm import Session
 from database.db import get_db
 from schemas.user import User, UserCreate, UserUpdate
-from database.crud.user import (
-    get_all_users,
-    get_user,
-    create_new_user,
-    get_user_by_email,
-    delete_user_by_id,
-    update_user_by_id,
-)
+from services.User.user_services import UserService
 
 user = APIRouter()
+user_service = UserService()
 
 
 @user.get("/users/", response_model=list[User], tags=["Users"])
 def get_users(db: Session = Depends(get_db)):
-    users = get_all_users(db)
+    users = user_service.get_all_users(db)
     return users
 
 
 @user.get("/users/{id}", response_model=User, tags=["Users"])
 def get_user_by_id(id: int, db: Session = Depends(get_db)):
-    user = get_user(db, id)
+    user = user_service.get_user(db, id)
     if user is None:
         raise HTTPException(status_code=404, detail="User not found")
     return user
@@ -30,15 +24,15 @@ def get_user_by_id(id: int, db: Session = Depends(get_db)):
 
 @user.post("/users/", response_model=User, tags=["Users"])
 def post_user(user: UserCreate, db: Session = Depends(get_db)):
-    new_user = get_user_by_email(db, user.email)
+    new_user = user_service.get_user_by_email(db, user.email)
     if new_user:
         raise HTTPException(status_code=400, detail="Email already registered")
-    return create_new_user(db, user)
+    return user_service.create_new_user(db, user)
 
 
 @user.put("/users/{id}", response_model=User, tags=["Users"])
 def update_user(id: int, user: UserUpdate, db: Session = Depends(get_db)):
-    is_updated = update_user_by_id(db, user, id)
+    is_updated = user_service.update_user_by_id(db, user, id)
     if not is_updated:
         raise HTTPException(status_code=404, detail="User not found")
     return is_updated
@@ -46,7 +40,7 @@ def update_user(id: int, user: UserUpdate, db: Session = Depends(get_db)):
 
 @user.delete("/users/{id}", tags=["Users"])
 def delete_user(id: int, db: Session = Depends(get_db)):
-    is_deleted = delete_user_by_id(db, id)
+    is_deleted = user_service.delete_user_by_id(db, id)
     if not is_deleted:
         raise HTTPException(status_code=404, detail="User not found")
     return is_deleted
