+++ schemas/auth.py
@@ -6,4 +6,3 @@ from schemas.debt import Debt
 class Login(BaseModel):
     username: str
     password: str
-
